refactor: Split EditSiteNoteScreen requirements logic from content
